Small test to avoid a useless overhead
When using component properties, two methods are injected in the instance (not in the class).
If the class has no properties, those two useless methods slow down the accesses to instance members.
If the class has properties, it may use the __slots__ special class member, containing the future getter and setter names (using the iPOPO constants modules) to avoid this overhead.