New method accept_binding() in servlets
accept_binding(path, parameters).
Called for each new servlet registration before the path availability
check.
If it returns False, the path check is not done, avoiding to raise
unnecessary exceptions.
Else, the path is checked (possibly raising a ValueError) and, if
available, the bound_to(path, parameters) method is called.
The servlet is finally registered if bound_to does not return False.