@Requires accepts only one specification
@Requires('field', ['a','b']) is not supported anymore, as it does not make much sense.
A requirement is for a (set of) service(s) that provides a specification.

The ipopo.core.Requirement has been updated to support only one specification.
This allows to use the latter as the first parameter of get_service_reference() (or alike),
and to test the filter only on services providing this specification.
=> This can speed the Simple/AggregateDependency handlers up to 5x