Added the immediate_rebind flag to ``@Requires``
Fixes #25
Avoids a component to be invalidated then revalidated if a service can immediately replace one that is being unbound.
The component is notified with the @Unbind* and @Bind* methods, like if it was losing an optional dependency.
If no service can be injected immediately, the component is however invalidated.

Example:
C depends on a service S provided by A and B, and A is initially injected in C.
When A is invalidated, B can now be immediately injected in C, without invalidating C.

C must have the following decoration:

  @Requires('field', S, immediate_rebind=True)