Implementation of the log service
Fixes #59

The OSGi's LogService and LogReaderService are implemented by the same class,
to ease the access to the entries store.
The LogEntry object is similar to OSGi's one.

The LogService keeps track of the standard logging entries, but doesn't handle
persistence: a LogListener should be implemented for that task.

The content of the log can be accessed using the "log" shell command.
Some lines can be added with the "debug", "info", "warn" and "error" commands.