Added pelix.framework.normalize_path()
This method ensures that all paths used for module import are absolute.
Both sys.path and the __path__ attribute of modules in sys.modules are
updated.
If the path isn't found, it is removed.

The method can be called before creating a framework, e.g. before changing
the working directory.
It is called by Framework.get_framework() before creating a framework
(this method is called by create_framework())

This should fix #61.