Converted LogService to a service factory
Each bundle gets a custom instance of the LogService, this avoids to
have to compute the calling bundle and should enhance performances.

The LogService and LogReaderService are now provided by two different
services, which causes an incompatible change.

The shell commands, the tests and the documentation have been
updated to reflect those changes.